requirements.txt
uvicorn
fastapi
tensorflow==2.16.1
keras
prometheus-client
#typing_extensions==4.0.0
Pillow
numpy
joblib==1.0.1
starlette